Move files SortedMap dir
