--- go.mod
@@ -1,3 +1,0 @@
-module github.com/Yuri364/golang_course_yuri364/sortedmap
-
-go 1.16
